requirements.txt
# Core
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.1
pydantic>=1.8.0

# Machine Learning
xgboost>=1.4.0
scikit-learn>=0.24.0
pandas>=1.3.0
numpy>=1.21.0

# NLP & Transformers
transformers>=4.10.0
torch>=1.9.0
tokenizers>=0.10.0

# Providers (Optional - using requests mostly, but google needs SDK)
google-generativeai>=0.3.0

# Environment
python-dotenv>=0.19.0
